requirements.txt

albumentations==1.3.1
chumpy==0.70
einops==0.8.1
hydra-core==1.3.2
imageio==2.37.0
numpy==1.23.5
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pillow==11.3.0
PyOpenGL==3.1.0
pyrender==0.1.45
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
smplx==0.1.28
tqdm==4.67.1
trimesh==4.8.3
xformers==0.0.28.post3
yacs==0.1.8
